Fix linux_spi.c build on 32bit systems.
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://flashrom.org/flashrom/trunk@1428 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ linux_spi.c
@@ -107,11 +107,11 @@ static int linux_spi_send_command(unsigned int writecnt, unsigned int readcnt,
 {
 	struct spi_ioc_transfer msg[2] = {
 		{
-			.tx_buf = (uint64_t)txbuf,
+			.tx_buf = (uint64_t)(ptrdiff_t)txbuf,
 			.len = writecnt,
 		},
 		{
-			.rx_buf = (uint64_t)rxbuf,
+			.rx_buf = (uint64_t)(ptrdiff_t)rxbuf,
 			.len = readcnt,
 		},
 	};
